--- gp_utils.h
@@ -1,19 +1,0 @@
-/* -------------------------------------------------------------------------
- *
- * pg_utils.h
- *
- * This code is utils for detecting active table for databases
- *
- * Copyright (C) 2013, PostgreSQL Global Development Group
- *
- *
- * -------------------------------------------------------------------------
- */
-#ifndef DISKQUOTA_GP_UTILS_H
-#define DISKQUOTA_GP_UTILS_H
-
-#include "storage/relfilenode.h"
-
-extern Size diskquota_get_relation_size_by_relfilenode(RelFileNode *rfn);
-
-#endif //DISKQUOTA_GP_UTILS_H
